source/requirements.txt
wheel
Discord.py
requests
pillow